+++ src/main/java/com/example/demo/exception/ResourceNotFoundException.java
@@ -9,9 +9,9 @@ public class ResourceNotFoundException extends RuntimeException{
     String resourceName;
     String fieldName;
     String message;
-    int fieldValue;
+    long fieldValue;
 
-    public ResourceNotFoundException(String resourceName, String fieldName, int fieldValue) {
+    public ResourceNotFoundException(String resourceName, String fieldName, long fieldValue) {
         super(String.format("%s with %s :%d  not found",resourceName,fieldName,fieldValue));
         this.resourceName = resourceName;
         this.fieldName = fieldName;
