+++ src/main/java/com/example/demo/repository/ISurveyAnswerRepo.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.SurveyAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ISurveyAnswerRepo extends JpaRepository<SurveyAnswer, Integer> {
+public interface ISurveyAnswerRepo extends JpaRepository<SurveyAnswer, Long> {
 }
